Fix Tab key handling.
The default CodeMirror behavior for Tab key handling always indents the current line. This fix inserts a tab character (or the appropriate number of spaces) if there is no selection, and indents the selected line(s) if there is a selection.

I thought this was a CodeMirror bug, so I logged it here:
https://github.com/marijnh/CodeMirror2/issues/328

Turns out this is expected behavior, so this is a change we need to make in Brackets.