JSLint integration.
Add new "Enable JSLint" menu item. When enabled, any time a .js/.htm/.html
file is opened or saved, JSLint is run on the file. JSLint errors are shown
in a table below the editor. If there are no JSLint errors, a gold star
is shown in the right side of the editor title area.

This checkin makes brackets.js and strings.js pass our lint standards. The
rest of our source files are not lint free.