Fix for issue #191 (UI left unusable if two dialogs are shown too close together in time) Bootstrap has numerous timing issues where dialogs break if one is shown while another's fade-out animation is still playing. We can avoid most of those by removing the "fade" CSS style that causes the animation from all our dialogs.
There's still one remaining Bootstrap bug, if a dialog is shown
synchronously after another one was hidden. The last notification that
Bootsrap sends out when closing a dialog is "hidden" - but that's not the
last task it performs. So anyone who responds to "hidden" by showing
another dialog will try to show it while Bootstrap is still in the midst of
hiding the last one. The fix is to reorder Bootstrap's hideModal() so that
dispatching "hidden" is truly the very last thing that happens.

Note: this still leaves issue #187 unfixed (showModalDialog() breaks if dialogs are ever stacked or shown sequentially too quickly).
The only way around that is to rearchitect our dialogs so that two
instances of the same type of dialog don't share the same chunk of the DOM.