First pass at auto-closing inline editor when Document is closed or is modified by something other than syncing from that inline).
Works, but is not super clean -- it might be nicer to have Document[Manager]
own the linking (mapping a set of inline editors to a shared Document)
and/or have Editor own the sync-guarding (distinguishing main editor changes
due to sync vs. main editor changes due to edit/refresh/etc.).

Also, consistent with the mini-spec this leaves a few edge cases broken: if
the file is not open when the inline is opened, we may miss external changes
and let the user silently overwrite them; or if the file is opened manually
after the inline was opened, we may overwrite changes made in the main editor
(becuase we don't start listening for main editor changes until the first edit
in the inline triggers us to link to the Document).