Tweak context menu event handling to support Windows context menu key. In general, anything using with mouse coords for purposes other than menu positioning should use mousedown, while the menu itself should be popped up via contextmenu.
It turns out that the special processing we were doing with mouse position
and selection is already done by CodeMirror too, so in this case we can
simply remove that logic instead of creating an additional mousedown handler.

This renders unused several Editor APIs that were recently added to support
the above logic. Removed one that seemed less useful, and cleaned up/renamed
the other two.

One caveat: when using the context menu key, menu position is not always
that nice. This is due to CodeMirror bugs: it doesn't always update the pos
of the focused textarea at the right times. However, in the pos of the menu
is often poor in other apps too, when opened via keyboard.