CodeHint changes: * Add CodeHintList.isOpen() method. Returns true if the code hint list is showing. * Add (and export) CodeHintManager.handleKeyEvent. This is now called from Editor.handleKeyEvent.
CodeHintManager.handleKeyEvent passes key events to the CodeHintList, if it is open. This avoids the need to add/remove event handlers whenever the hint list is opened.

There is no reliable way to know when the CodeHintList is closed, so instead we do something less savory. When the list is opened, an "opened" flag is set. After that, any time isOpen() is called, if the "opened" flag is still true, we check to see if the "open" class is still attached. If not, we clear the "opened" flag. This adds a bit of overhead while the list is open, but should not impact performance at all when the list is closed.