Pop up attribute value code hints immediately after committing attribute name hint.
As part of this, made it so CodeHintList is no longer a singleton within
CodeHintManager--each new popup is a new CodeHintList. This is so we don't
have issues with order of operations, where something tries to close the
previous CodeHintList after a new one has popped up. Now a CodeHintList
will only try to close itself (and it's okay if it tries to do it more than
once).

Also included a partial fix for #1381.