Fix bug #1627 ("Go to Definition" prepopulated text doesn't work when activated via menu) -- There are two problems at play. First, smart-autocomplete only listens for key events, not programmatic or other textfield changes, so we need to kick it to generate results for the prepopulated text.
Second, the results are generated asynchronously while QuickOpen's auto-"focus"
is synchronous. That bug is not fixed here since it's related to a whole
tangle of async issues in #1855.

Fixing the first bug alone means that hitting Enter now goes to the correct
result (instead of doing nothing), but that the editor doesn't auto-scroll
to that result as soon as the search bar is opened.

This commit also cleans up QuickOpen to use KeyEvent consts for key codes.