Focus changes
EditorManager.focusEditor() now returns focus to the last visible
editor that had focus. We no longer force the focus back to the
main editor after saving or cancelling an open.

Add EditorManager.getActiveEditor(). This function is similar to
getFocusedEditor(), but will return the last editor that was given
focus, even if it doesn't have focus at the moment. This is useful
for commands like Find/Find Next that should work on the active
document even if the Find dialog has focus.

Update FindReplace, FindInFiles and QuickOpen to use getActiveEditor()