Partial fix for #2015 (Bottom panels get messed up if resized to max height).
When panel size > the available area (which can happen several ways,
including a bug in Resizer that allows drags to overshoot the max size that
would fit), we were computing a negative size for the #editor-holder. This
was a problem because jQuery no-ops on negative sizes; clipping to 0 gives
much better results.

#2015 should remain open to track the other panel-oversize-related bugs
mentioned there.