fix 2068: rework quickopen logic to produce better results
Created a new QuickOpen matching algorithm that searches for matches
among "special characters" (path markers, camelCase changes) and
does so left-to-right rather than right-to-left as in the old algorithm.
The new algorithm does still give an added bonus to matches that occur
within the filename, which is checked first.

Also, there is now a collection of tests to try out the QuickOpen logic
and ensure that it is working sanely.

The scores look pretty sane now.

In this commit, I added some code to help debug scores and was
able to get the scores to be pretty sane and the results look quite
nice.

I also removed the dead code.

fix a display bug when there is no query

added comments for new QuickOpen algorithm.

Note that in the process I spotted a bug and added a failing test,
which I have not had a chance to fix yet.

partial fix for a bug in the new quickopen logic

fixed the bug with strings that are longer than the final segment

fix an off-by-one problem that left an initial character match out of
the last segment