fix _computerangesa not matching _computeRangesAndScore.
The problem was that backtrackTo was causing backtracking to go
too far back for the "s", because it had already backtracked to
the "r" previously (when it hit the "g" in the query).
backtrackTo was the original mechanism I used in backtracking
before adding deadBranches. It turns out that backtracking
really needs to go back before deadBranches[queryCounter],
because where we need to backtrack to depends on where we are
in the query.