Better error handling when installing updates.
Update installation should generally not fail given the checks that
happen on the first pass of installation, but if they do we have
a reasonable display for the user and a way to get at the root
error.

Also cleaned up some of the extra braces in the extension manager
strings.