reorganize Readme
Moved paragraphs around so least technical (most user-oriented) is on top, then tech/contributing (hacker info) is near the end.  This allowed to streamline the prose, but I tried to avoid changing the text as much as possible.

Linked blog in welcome section (great for finding features; helps users decide if Brackets is for them).

Used H4 vs **bold** for sub-headers.