When renaming a file, merely changing case should not produce an error that a file preexists with that name.
This commit enables renaming a given file named, for example, "file" to
the same name but of a different case (e.g. "File"). Without this
commit, such a rename attempt will erroneously produce an error that a
file already preexists with that name.

The true fault lies with the low level method brackets.fs.rename, which
does not permit renaming a file in this manner for some reason. This
commit is designed to work around that limitation at a higher level.

A perhaps better solution than this commit down the road will be to
upgrade the fs module. However, for the time being this commit fixes
this single bug in a more isolated way without requiring any global
changes.