Implement document proxy icon in Brackets editor
This change requires request
https://github.com/adobe/brackets-shell/pull/322 to be
merged.

On Mac, application name is not usually shown in window title.
Instead a "document proxy icon" is shown as a shortcut for file access.
This icon can be dragged to other applications (e.g. Mail or FTP
client) or used to open containing folder in Finder.