Implement document proxy icon in Brackets editor
This change requires request adobe/brackets-shell#322 to be merged in.

On Mac, application name is not usually shown in window title.
Instead a "document proxy icon" is shown as a shortcut for file access.
This icon can be dragged to other applications (e.g. Mail or FTP
client) or used to open containing folder in Finder.