Standardize back on trailing-/ for directory paths. This is safer for legacy code that concats paths with this assumption. The normalization hook in FileSystem also removes duplicated /es. In the future it could also handle "/." and "/..", but it doesn't yet.
Fixes the infinite-traversal bug in JS code hints & re-enables the unit
tests that were hitting it. But adds some hacky temp code to forestall race
conditions when running these tests (presumably a race on master too).

Many of the diffs are just restoring code back to what it is on master.