Set preferences are, by default, set in the context in which they are defined.
For example, if a preference is set at the "user" level and the user changes
the value via UI, the preference will be changed at the "user" level. If it's
set in a .brackets.json file, the preference will be changed there.
If the preference is defined only at the "default" level, it will be set
at the next higher level of precedence ("user" level in our setup).

This is a fix for #6595.