Line drawn here (implementation change to come).
This was the starting point of a refactoring to better handle a larger, disjoint
set of path-based Scopes. I can see complexity brewing and I don't think it's
worth it for path-based Scopes. After this commit, I'm going to yank out all of
the path-based Scopes code and switch to simpler, project-based approach.