Implement jscodehints.defaultExclusions preference.
We've had a number of people report problems with Ionic in their projects.
Tern appears to have an issue with the minified Ionic files. I had planned
to add default exclusions so that we could quickly solve pain like this
and this problem appeared to be a good candidate. The repro steps
[reported here](https://github.com/adobe/brackets/issues/7262#issuecomment-41851396)
are fixed with this change.