Signal change in file tree when selectionViewInfo changes.
There is a special case in which we don't signal that change (it's noted in the
code). This should eliminate some cases of inconsistency. A related change is
that setSelection will no longer select items that aren't visible in the tree
(something that does not happen in real life anyhow).

This is a fix for #9343.