Restore the QuickOpen.basicMatchSort and friends.
This change broke a few extensions. Looking at the fix, it seems that all QuickOpen plugins will
likely need these functions, so we may as well re-export them as we have been doing rather than
requiring the extensions to all add another import.

If we do decide to deprecate these later, we should do so with deprecation warnings (something we
weren't doing when these were moved to the StringMatch module).

Revert "Marked as deprecated by adobe/brackets#2462 in Sprint 19"

This reverts commit 49e08274a0bc9259d6dd94c03351e4c87bf570f9.