Workaround for #10124: change the title on close.
On Mac/OSX alert causes system modal dialog which prevents any interaction with
the browser and navigating away. This was causing unit tests to fail.

So instead of showing alert, change the title of the document to indicate that
the Live Preview session was closed. A more clear UI/UX definition is needed
for definitive solution.