Removed outdated comments.
Also, show the malformed message when such message comes in. Removed the TODO
comment wrt sending back the error message: it is unclear what should be done
if a malformed message comes in. It is unlikely it will ever happen and even
then the error message in the browser should be enough.