Fix for #11763 The new created file cannot be renamed/showed in file tree/showed in explorer with no messages if it isn't saved.
The file is checked if it is an InMemoryFile before attempting to perform
the above mentioned operations on it. If it is indeed not saved, then a
message is shown to the user to save the document to proceed. Upon
clicking 'Ok', the save dialog box opens up allowing the user to save the
file.