Browser issue (#12946)
* Browser issue fix

These two sections check for Urls being dragged in to the browser.

* There was a bug in the unit test

So for some reason this changed which I did not perform showed up in my
master it seemed that one did not have access and the new one did. In
the health status test under extension tests, the connection would
timeout.

* JSLint syntax fix

JSLint want’s this instead of “==“. I had actually done a check with an
online checker and I forgot to make the change.

* Revert "There was a bug in the unit test"

This reverts commit 887e3d5045964da383959fbe5218d7393bd6879a.

* Simplify URI-list antipropagation

These are the changes as requested by @petetnt. The behavior was packed
into a function. Generally `types` is not null but was protected on the
case it is. Files must be null or the length must be 0 now.

* Indentation and spacing

Fixed extra indentation and spacing issues

* Clean comments

Cleaned up the comments, some were redundant to the functions
description. I gave the different cases of `types` values when files
and text is dragged/dropped into Brackets.