Fixes wrong function call in ProjectManager.js
The createNewItem function in ProjectManager.js calls the createAtPath in Projectmodel.js.

The function call expects a single parameter, which is the path at which the file must be created.
And it expects a "/" in the end of the path in case it is a folder.

However, right now 2 parameters are being passed. One of them being the isFolder variable.